+++ tests/test_constants.py
@@ -549,7 +549,7 @@ def test_serial_comm():
     assert STIM_SHORT_CIRCUIT_THRESHOLD_OHMS == 10
 
     assert VALID_STIMULATION_TYPES == frozenset(["C", "V"])
-    assert VALID_SUBPROTOCOL_TYPES == frozenset(["delay", "monophasic", "biphasic"])
+    assert VALID_SUBPROTOCOL_TYPES == frozenset(["delay", "monophasic", "biphasic", "loop"])
 
     assert issubclass(StimulatorCircuitStatuses, IntEnum) is True
     assert StimulatorCircuitStatuses.CALCULATING == -1